added original image file
This file is what can be used with the provided output statement. In the source I simply pasted the image into a 'define' and named it 'beach'.
I wasn't sure if the file would be saved properly in the .rkt file so I just figured I would attach it here as well.