added new encrypy.rkt
now the line with the picture definition has a dummy line with a comment detailing how to run the program normally.  Copying the beach.png image from the repository and pasting it on the arrowed line like so '(define beach [pasted image])' will allow the program to run as intended.